Add ERR, ERR1 and ERR2 Measure Functions
This commit adds the ERR, ERR1 and ERR2 measure types for AC,
DC and TRAN measure modes.  Note that the ERR and ERR1 measures use
the same Err1 class.  Also, note that functions common to the Err1
and Err2 classes have been placed into the ErrorFunctions class.
So, these three measures have a slightly different class structure
than the other measure types in Xyce.

The default values for weight_, ymin_ and ymax_ were changed,
in the Base object, to conform to the HSPICE defaults.  Also,
ymin_ must now be non-negative, while ymax_ must be positive.
Finally, IGNOR was added as a synonym for IGNORE.

SON Bug 1278
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1278