Rename RelativeError Class for ERROR Measure
The RelativeError class for the ERROR measure did not, in fact,
calculate a relative error between two inputs.  So, this commit
changes the name of that class to Error, which seems like a more
appropriate name.  Note that the ERROR measure is different from
the ERR, ERR1 and ERR2 measures.

This commit also fixes a cut-n-paste typo in the EquationEvalution
class for the EQN measure.

SON Bug 1278
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1278