Improve the resolveExpression function
With this commit, the newExpression::resolveExpression function has been
modified so that it attempts to resolve all external expressions that it
depends upon, and not just the local one.  This will make it easier to
do resolutions of expressions which depend upon nested sets of
functions.

The motivation for this was that I wasn't sure in the Xyce source where
would be a safe place to force all the expressions to undergo a
resolution.    In my original test codes, there were very clear setup
phases, and it expected that all expressions would complete one phase
before moving on to the next one.  So there is one initial phase that
does the lexing and parsing and all the initial allocations.

Then there is a second phase were external dependencies are checked and
resolved.  This is done by attaching relevant nodes to paramOp and
funcOp classes.   But this can only be done if the first phase is 100%
complete.  Otherwise it will miss stuff.

So, in my unit tests, it was easy enough to ensure that this happened.
But in the Xyce source, much less clear as expressions are used all over
the place.  But, presumably the implicit phases are distinct enough for
the old expression library to work.   But I didn't really want to add a
new layer of resolve expresison calls to Xyce.

So, I changed the newExpression::evaluate and
newExpression::evaluateFunction functions so that they would check if
they've been previously resolved, and if not, call "resolveExpression"
before evaluating.  This worked fine, except that the
"resolveExpression" of  expression A didn't force the
"resolveExpression" of external dependency B to be called.

So, I've updated the resolveExpression function to call the
resolveExpression function on all subordinate dependencies.  That way,
as long as they've all been parsed, they should resolve sufficiently
before being evaluated.

I have added a unit test of this capability, called
TEST (Double_Parser_NestedFunc_Test, func_cir)

The other main change in this commit is that the N_UTL_Expression
library now stores the newExpression pointer as an RCP.  This forced a
number of other changes, particularly to the group.   As the
expressions are copied all over the place, this seemed a safe way to go.