Make sure not to bother computing unrequested derivatives
Make sure that we don't bother working out derivatives when they
haven't been requested.  Specifically, don't work out the
derivatives with respect to "derivProbe2" and second derivatives
if "derivProbe2" isn't set.

All the low level templates (where the real work is done) were already
doing that, and were returning 0.0 for any unrequested derivs, which
should be good enough to make sure that higher level stuff also
returns 0.0, but it is better not to bother at all when possible.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410