Remove unnecessary parens and unary "+"
Xyce/ADMS takes a lot of pain to reduce derivative calculations if,
for example, it finds that it wants to multiply something by 0.0, or
by 1.0.

1.0 or 0.0 are often returned as the derivatives of things, and as we
construct expressions going up the tree it is important to preserve
them, so that later on we can recognize them as simplifying values.

This commit removes some wholly unnecessary parens and unary "+" signs
that do nothing more than take a derivative value returned from
elsewhere and add meaningless decorations that would later interfere
with further simplification.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410