Add a template to handle second derivs of two-argument builtins
In obscure cases, when ddx() is used to set the value of a variable
which is ultimately used to compute the RHS of a contribution, we need
second derivatives (the first derivative is the ddx() value, and
the second derivative because we need the first derivative of the
ddx() expression for the jacobian).

This ultimately means that if the expression in the ddx() (or any of
the expressions for variables that are used to obtain the expression
in the ddx()) contains function calls, we need to be able to
differentiate those calls.

This commit creates a template that, given a function call, returns
the second partial derivatives of the function with respect to its
arguments.  e.g. if given f(x,y) it returns d^2f/dx^2, d^2f/dxdy,
and d^2f/dy^2.

This is strictly symbolic differentiation.

These partials with respect to arguments will ultimately be used along
with derivatives of the arguments themselves to build up the total
derivative of the function call with respect to two unknowns.

Interestingly, when I look over other ADMS back-ends, they only do
this for pow(), and only when the second argument is not dependent on
probes.  In that case, you only need d^2f/dx^2.

I'm also only going to do pow(), but create infrastructure that can be
extended to other functions in the future.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410