Add more 2-var functions, additional deriv
This commit adds support for second derivatives of min, max, and hypot
to the twoArgSecondDerivs template (which, as yet, is not called
anywhere).  min/max are easy --- their second derivatives with respect
to their arguments are zero.  hypot is messy, but not very.

I also added the code to compute the first derivatives with respect to
"derivProbe2" of two-argument functions that are already supported
for derivatives with respect to derivProbe.  These are also needed for
ddx() support.

I still don't have anything wired in to use twoArgSecondDerivs, but
I'll get there.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410