Fix syntax errors, add function 2nd derivs
This commit adds code to the "function" template to actually compute
second derivatives of 2-argument builtin functions (pow, min, max,
hypot), and first derivatives of analog functions with respect to a
second probe (needed for ddx()).

Second derivatives are only needed for ddx() when the result of the
ddx() is used in a contribution.  These are not implemented for analog
functions.

Also fixed in this commit are a few syntax errors that choked admsXml.

As it stands with this commit, this back end runs correctly on the
existing models I tested it on (vbic, bsimcmg) and will do nothing
differently than it did on the main branch except for the absense
of unnecessary parentheses that were removed in an earlier commit on
this branch.

It will still do nothing with the extra derivatives it's set up to
calculate, because I still haven't actually written the code
to *invoke* that stuff.  But this is a significant checkpoint, because
the back end is actually running right now and is producing correct
results when the model does NOT  use ddx().

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410