More use of RCP in newExpression/N_UTL_Expression
This commit also includes some cleanup.  It also represents the final
commit on this branch before I gave up on attempting to re-purpose the
old API with the new expression library.  The dividing line isn't 100%
black and white; this commit includes a few things that add (optionally)
to the old API.  But for the most part, the really big refactor to
create and use a completely new API is happening on another branch now.

I've been staring at code a LOT, and not writing that much this week.
I've been attempting to figure out how to best get nested functions to
work correctly.   Expanding the use of RCP makes this a little easier I
think.

One idea I had is to make the xyce group aware of objects from the
circuit context such as:

    const Util::ParamMap & context_function_map = output_manager.getMainContextFunctionMap();
    const Util::ParamMap & context_param_map = output_manager.getMainContextParamMap();
    const Util::ParamMap & context_global_param_map = output_manager.getMainContextGlobalParamMap();

These are used in the old expression usage a bunch anyway.  So I thought
maybe it would be easy to do that with the new way; just pass these maps
into the group, and when stuff gets resolved, just look them up, and
grab the needed AST node and attach it to the present expression.

So, I started writing code to try this out and as a result added some
features to the xyceExpressionGroup class.  This unfortunately required
that the expression group be made aware of the Xyce::Util::Param class,
which it had been unaware of before.  This expanded the dependencies in
that file more than I would have liked.

Also, this may still work, but it is a little harder than I thought it might
be, as these objects (as they come from the OuptutMgr) are all const.
And the output manager receives them from the circuit context
(indirectly, via N_CIR_Xyce) as const.  I think I probably need to do
things with them that would require they not be const.  So...  I may
need to find another way to set this up.

I probably need to create a proper Xyce group class, that is set
up well upstream from most expressions.  It will need to get these maps
directly from the context class, or possibly be set up within that
class.