Enable Xyce/ADMS to emit "required" second derivatives
This commit enables Xyce/ADMS to begin to emit second derivatives that
are required by the presence of ddx().  Sorta.

I say "sorta" because sometimes they are NOT required, because the
quantity that is computed using ddx() is only used in noise
contributions.  Xyce's noise code does not require the derivatives,
and therefore does not require the second derivatives that result from
using ddx().  ADMS does not distinquish between quantities used only in
noise contributions and those in other (static or dynamic)
contributions.

As a result, in its current state this code might emit derivatives
expressions that involve unimplemented analog function derivatives,
but which are ultimately not used.  At the moment, those won't compile
because I deliberately have the back-end emitting syntax errors in
those cases.

This is still not emitting code that implements ddx(), but now that
it's in the current state there are only a few things left to do to
get it done.

This back-end has been tested by running it over all existing
Verilog-A models to make sure it doesn't break them, and also by
running it over a hacked version of the "diode2.va" toy that includes
a ddx() call.  It emits all the derivatives needed to compute the
ddx() and anything downstream of it.  It just doesn't emit the actual
code for ddx().

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410