Regenerate all models derived from Verilog-A
In my bug 410 branch I've modified the back-end to emit fewer
unnecessary parentheses.  Now that I've got a version of the back end
that runs and is almost ready for the next phase of ddx() hackage, I
want to save the output of this back-end so that I don't have to wade
through tons of diffs that differ only by parens, and can see clearly
what the difference is between versions I generate and scan for
erroneous output.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410