Fix some parenthesis problems in second derivatives
If a second derivative of a subexpression turns out to be somkething
like "a*b+c*d", we had better return that subexpressions derivative as
"(a*b+c*d)" or there will be problems when it turns out that the
subexpression was in this context:  e*subexpression.  As, for example:
  h=e*f*g
where e is a constant, the correct derivative should be:
  d_h_dV... = e*(d_f_dV...*g + f*d_g_dV...)

Prior to this commit, the parens would have been missing on the RHS.

I think I have gotten all the spots where this was a problem, but will
have to do some heavy testing to catch all the possibilities.  This
error was not picked up in the tiny toy test I had.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410