Fix uninitialized variable in ADMS template
The "argsDep2" variable was not set at the level of context where it
was used, and therefore all the settings of that variable at lower
level were local to the lower levels.  This didn't matter in any of my
testing because that block of code wasn't exercised.  It was
exercised when I tried to look at the BSIM4 SOI.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=203
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410