Remove debugging text, fix missing space issue
I had put in a bit of code yesterday to add a comment after each
variable declaration saying how that variable had
"insource" set (i.e. used in a chain of computations that impact the
RHS of a contribution (source) or not).  It was not meant to be kept.

Also, I find that our "$strobe" code was not correctly putting spaces
between strobe output arguments, so they all ran together.  BSIM likes
$strobe, and uses it for all sorts of informational output.  Xyce/ADMS
does $strobe sloppily, but this makes it at least slightly readable.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=203
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410