Modify diff ordering in ADMS scripts
generate_ADMS.sh regenerates all ADMS-generated models from the
current Xyce/ADMS backend and then runs a diff on the output against
the old versions.  compare_ADMS.sh just does the compare of the
versions in the ADMS/examples directories to those in the
src/DeviceModelPKG/ADMS directory.

When D.B. originally created these scripts, he made the diff
compare the new to the old, so what diff showed is how to patch the
new back into the old.  It made for some confusing output for me,
because I'm used to diff answering the question "how did the new one
change from the old?" not "how would you change the new back into the
old?"

This commit reverses the order so that the diff output shows how the
old morphed into the new, not how to morph it back.