Fix erroneous diff lines for ekv2.6
I recently changed generate_ADMS.sh and compare_ADMS.sh to tweak the
ordering of diff arguments, and did so using an editing macro on every
line of the file.  I failed to notice at the time that when I added
ekv2.6 to the script a long time ago that I had already done it in my
preferred order, and so reversed it.

This fixes that mistake.