Make Xyce/ADMS generate code to handle output vars
The Verilog-A standard says that any variable that has "desc" or
"units" attributes is supposed to be an output variable --- i.e., one
that should be output with .OP.

Xyce doesn't do .OP in all its glory (yet), but does allow output of
internal variables using the N() operator on .print lines.  Doing so
requires that the code for a device have extra stuff in it to put the
values of the output variables into the store vector, and to give
those locations in the store vector a name.  That name is generally
<device instance name>:<variable name>, output on the print line as
N(<device instance>:<variable>).

This commit adds the necessary stuff to the non-Sacado Xyce/ADMS
back-end (the only one we're going to support going forward) so that
this type of output variable designation actually works.  This
requires that:

  -- adms.implicit.xml be modified slightly so that it
     actually flags variables with desc or units attributes as output
     variables.  The original version shipped with ADMS did not do so,
     because the part of the standard that created this convention was
     written long after ADMS development had matured.  Prior to that,
     an ADMS-specific attribute called "ask" was the one you used to
     tell ADMS to treat it as an output variable.  That feature still
     exists, but is non-standard.

  -- Code be added to declare LIDs for the store vector to hold output
     variable values.

  -- Code be added to define node symbols for the output variables.

This commit contains only the ADMS back-end changes needed to
implement the feature, and changes to a few Makefiles and patch files
to use it (one makefile was improperly still using the old back-end
header generator(!)) and others just needed changes to their clean
targets.

This commit also fixes a couple of long-standing omissions.  Turns out
that I had never been including store LIDs in initializers of the
constructor (they're also needed for voltage limiting), and I was
never initializing collapse booleans in the constructor, either.
Neither thing mattered, because these variables are never used before
some other bit of code assigns to them, but now Xyce/ADMS emits them
for completeness.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788