Enable output vars for both hicum devices
My previous commit made grander claims about what output variables
hicum had.

Some of hicum's output variables used the old-style ADMS designation
of setting 'ask="yes"' for the variable's attributes, and others used
the new-style Verilog-A standard of setting "desc" or "units" to
anything non-null.  The latter were ifdeffed out, and the former were
not.

This commit unifdef's the ones that were ifdefed.  Some of them are
now inside a new ifdef, because they use a Verilog-A construct we
don't support yet, "port currents".  That is, "I(<a>)" is the current
through port "a" of the model (i.e. its lead current).  We compute
lead currents, so such a thing would be possible, but it's not quite
in a state where it can be done easily and correctly.  So I just
ifdeffed out any output variable that depended on them.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788