Fix erroneous ddx generation for output vars
The ddx() generation was set up so that it only handled "insource"
variables correctly.  When ddx() is used for a variable that is only
used for output, it was improperly generating "0.0".

This commit fixes that error, which at the moment only broke a couple
of output vars for hicum_l0.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410