Fix a number of errors with ddx()
While working on testing ddx() under the auspices of bug 788, I found
a number of errors in the handling of ddx() for output variables and
for the case when it turns out the expression argument of ddx()
doesn't depend on the probe with respect to which ddx() is
differentiating.

While I was at it, I fixed a couple of cases where mapply_binary was
not detecting 0.0 arguments and using unnecessary parens.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410