Enable BSIM6 output variables
Though I had earlier claimed that BSIM6 output variables were Just
Working, I hadn't tested them.  Turns out that BSIM6 output variables
were *DECLARED* and could therefore be printed on a .print line, but
the code that actually *evaluated* them was ifdef'd out.

When I unifdefed them, it turned out that there was some breakage.
Some errors were still present in ddx() handling, and some usage of
ddx() in the model was currently unsupported in Xyce/ADMS.

The errors in ddx() were fixed on the feature-bug410son branch, and
I've hacked on BSIM6 so that all the ddx() usage that was unsupported
is recast into equivalent forms that *are* supported.

I have a test for this, and will commit that to Xyce_Regression
shortly.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=410