Make PSP102 output variables actually work
As I had done for several other models, when I added PSP102 to Xyce
I'd put an ifdef around all its computation of output variables,
because we didn't support outputting them, and that made it wasteful
to compute them.  But I didn't comment out their declarations.

Now that we DO support output variables (those that have "desc" or
"units" attributes in the Verilog-A source), we were set up to output
the variables, but they were always zero.

This commit removes the ifdef I put in, and also enables the OPderiv
macro that the PSP authors had put in there to control whether or not
to try to ddx() things --- apparently some other simulators besides
Xyce had trouble with ddx(), too.

There is a test associated with this commit in the Netlists/PSP102
directory.  It follows a pattern similar to the ones used to test
BSIM-CMG, PSP103, and BSIM6 output variables.