Improve unit test coverage
Now that I can run the unit tests thru gcov/lcov, I was able to see what
features I was not testing.  This commit is a quick cut at fixing a
bunch of the low-hanging fruit.  This includes testing the "output" and
"codeGen" functions (nominally).  It also includes a bunch of operators
I had neglected to test.  And, it includes one bugfix for the "or"
operator, which wasn't being parsed properly.

The ast.h file is at this point only 60% covered (roughly) by the
parserUnitTest program.  Before today, it was around 30%.  So that is a
big improvement.   A number of the other operator headers are (now) at
100%.