Add BSIM-CMG version 110 output variables
This commit enables the output variables originally defined by the
BSIM-CMG authors.

Due to an early user request for some output variables for this
device, we had previously added "VDSAT", VDS, VGS, and VBS output
variables.  None of these variables are in the set defined by the
original source code.  VDSAT is in fact a duplicate of what the
authors called "VDSSAT", and the others are just voltage drops across
the external nodes.  These variables have been added back in to the
Verilog-A source.

This feature was already being tested, and the tests continue to pass
even though the method of supporting the feature has changed.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788