Add output variables for MEXTRAM
This tooks some rather heavy-handed changes, because the MEXTRAM
source was a bit muddled as to when it did these calculations (they
ifdeffed it with something more generic, VAMS_COMPACT_MODELING), and
had a broken OPP macro (to define operating point variables) that
incorrectly took unquoted arguments and tried to quote them --- the
result was that every variable had a description of "des" and units of
"uni".

It is also difficult to test, because like hicum, they have a
transconductance output gm, but don't actually output the thing of
which gm is the derivative.  So it's not possible to compare it to a
finite difference approximation.

I have manually run tests on it and seen that it is outputting
information and not just all zeros.  Since it is pretty well
established at this point that the output variable mechanism is
working correctly and ddx() is properly performing symbolic
differentiation, I don't feel it is necessary to test every single
output variable for correctness.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788