Enable MVSG-HV output variables
MVSG-HV is, as yet, undocumented, untested, and generally unused in
Xyce.  Still, enabling its output variables was a simple matter of
removing an ifdef I had placed in there to omit them.

I've added the output vars table to the reference guide directory, but
as yet the model is undocumented in the reference guide, so this table
is not being used.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788