Support DERIV, INTEG and RMS Measures for .AC and .DC
This commit adds support for the DERIV, INTEG and RMS measures for
.AC and .DC measures.  For the DERIV measure, the RISE/FALL/CROSS
qualifiers are not supported for AC or DC measures.

The recent changes to how the WHEN time is interpolated, for the
FIND-WHEN measure, were added to the DERIV measure for AC and DC
measures.  Those changes still need to done for TRAN mode for
the DERIV measure.

This commit also changed how the RMS measure values are calculated for
AC, DC and TRAN measures.  The new approach follows the HSPICE
definition of "the square root of the area under the out_var curve,
divided by the period of interest."

For the AVG and RMS measures, for AC and DC measure modes, a
measure with only one point (e.g., if FROM=TO) is a FAILED measure
by definition even though the FROM-TO window may be valid.  This
commit adds printMeasureWindow() functions to the Average and RMS
classes so that the measure window info, printed to stdout, is
correct for those corner-cases.

SON Bugs 1282 and 1283
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1282
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1283