Make Xyce/ADMS skip emission of useless assignments
Sometimes, in the course of processing a Verilog-A model, Xyce/ADMS
would emit assignments like:
   A = A;

Most of the time this was the result of expressions that had some sort
of conditional compilation wrapping them, like an ifdef, but it is
happening more now that Xyce/ADMS is emitting derivative expressions.

This commit simply makes Xyce/ADMS skip emitting an assignment if the
LHS and RHS it would have emitted are identical strings.

This shuts up a lot of warnings from Clang.