Restore hspice "and" and "or" and test
In a recent commit I commented out the Hspice style "and" and "or"
tokens from the lex file.  (i.e., I disabled "&&" and "||").

I did this because I discovered (via gcov) that I wasn't testing most of
the comparision operators, and when I started testing all of them, it
turned out that the "or" operator was broken.

I wasn't sure at first why it was broken, so to be safe I commented out
the alternate Hspice forms of these operators, leaving only the Xyce
versions. ("&" and "|").

That didn't fix it.  It turned out that I had incorrectly included "|"
in the list of single-character tokens near the top of the lex, file,
and that meant it had precedence over the later TOK_OR.   So, the fix
was to remove it from the single-character list.   I did that, but then
didn't restore the Hspice style operators.

Today, I did restore them and I added tests for them.  They pass, so
the Xyce and Hspice style AND/OR operators can co-exist in this parser.

Also, in this commit, I have changed some of the google test "EXPECT_EQ"
invocations to "EXPECT_NEAR".  A number of the single character ternary
operators (which were producing many digit fractional floating point
numbers) were failing when I switched from clang to gcc.  EXPECT_NEAR
allows you to set a tolerance.  Unfortunatley I had to make the
tolerance a little larger than I would have liked but now most of them
pass.

For all of these tests I was already subtracting the expected answer
from the real answer, and then comparing the result of that subtraction
to zero.  So, when they failed, I could see exactly how much they were
failing by.   Based on this I set the tolerance to 1.0e-15, and this
fixed all but two of them.