Clean up zero detection in mapply_binary template
The Xyce/ADMS template for "mapply_binary" (binary operator) is one of
the oldest in Xyce/ADMS, and was patterned on some other ADMS
back-end's template that I found back in 2009.  It tries to find
places where it can leave out terms and factors by checking if they're
zero or one, but incorrectly does so by looking at the math/value bit
of the ADMS tree instead of the processed argument string--- that means it
only finds expressions that are identically 0.0 or 1.0  in the source,
and that if a subexpression evaluates down to 0.0 and *then* is used
as an argument to a binary operator, it might get missed.

This commit, which has no impact on any existing Verilog-A model
currently in Xyce, augments all of these tests to look at the
processed string for an argument as well as its numerical value.  In
some future model, we might find that it throws out some extra terms
at some point.