Fix glaring but inconsequential error in Xyce/ADMS
A block of code that was supposed to be inside a conditional was not.

Didn't matter, though, because what that block did was ignored if
the conditional wasn't true, but there was no point doing the work
if it was going to be ignored.

This was code to evaluate a term of second derivatives in the case
where one was dealing with a division operator.  It should not be
getting executed at all if one isn't asking for them, and it was.
And then all the strings it created were being ignored (because they
hadn't been asked for).

Has no impact on any Verilog-A models we use, but was just wasteful
processing.