Add interpolation to short circuit currents in S-param model
The S-parameter model in frequency domain will be extended to add short circuit
currents in order to model the circuit response to EM environments. It is
necessary to interpolate the short circuit currents for all ports if the data
given is not at the frequencies for HB analysis. This commit adds the
interpolation for short circuit currents in the model.