Begin work on removing Sacado from sensitivity
Xyce compilation times have skyrocketed since the implementation of
Xyce/ADMS improvements for bug 410 (DDX support) and bug 788 (support
of output variables).  That's because supporting and enabling output
variables has ballooned the contents of the various
sensitivity-related functions, and hence inflated the amount of Sacado
usage in these functions.

During work on ddx support, I realized that it would not be an
enormous amount of extra effort to replace Sacado use in
sensitivity *entirely* by having Xyce/ADMS emit the required
derivatives.  This should have dramatic effect on improving the
compilation time (and memory requirements) of Xyce in the ADMS
directory.  The issue related to this work is bug 1284.

This commit gets us part of the way.  It augments most templates that
evaluate terms of expressions, and adds to them the ability to compute
their derivatives with respect to an unknown "X" as well as their
derivatives with respect to probes for normal context or ddx()
requirements.  X will eventually be some instance or
model parameter, but that will be determined at run time.

It is basically an exercise in cutting and pasting the code that
generates those other derivatives, with minor modifications so they
only apply to expressions that have parameter dependence instead of
probe dependence.

I am not committing this because it actually does anything useful.
I'm committing it because it's currently in a state that doesn't break
anything.  It does not yet *emit* the derivatives, but it is working
them out internally.  With this commit, it still produces *identical*
output to what it did before.  The point of this commit is to save a
working state so I don't break anything in the next step and wind up
taking two steps back.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1284