Make Xyce/ADMS initialize all variables on declaration
Turns out that some variables in sensitivity were being declared but
never set.

These were mostly derivative variables that were associated with
variables set with the value of limited probes.  They were tagged as
being parameter dependent (because of the limiting), but in
sensitivity context the limiting doesn't actually happen (because it's
done on converged solutions) --- thus, the actual dependence on
parameters should be zero, but were never being set as such.

This commit changes the template that does variable declaration
to *always* initialize the value to zero.

This was the source of my NaNs in VBIC sensitivities.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1284