Regenerate VBIC code with non-sacado sensitivity
This time for sure.  The previous version had failed to initialize
some _dX variables, and the result was bad.  This version is generated
with a back-end that *always* initializes variables on declaration.

This time, it actually produces correct values for the sensitivities,
not just "passes broken xyce_verify" testing.  (xyce_verify declared
the VBIC sensitivity tests passing even though the "test" file was
full of NaNs --- scary, but a *different* bug)

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1284