Extend S-parameter device to include short circuit currents
The standard S-parameter model is passive. It can model any passive linear
network and is widely used in many applications. However, it cannot
model the linear network with current/voltage sources. The S-paremter
model is extended to characterize linear networks with sources. The
linear system is reduced to a Norton equivalent circuit. This includes
Y parameters in parallel with short circuit currents. The extended
S-parameter model is tested with the data from an EM solver. The results
match the gold standard from the EM solver. The extended model is
supported for frequency domain HB analysis