Reduce size of sensitivity function argument lists
Turns out that a HUGE contributor to the compile time issue of
sensitivity is the ENORMOUS argument lists that some of them had.

That's because all model and instance parameters and all model- and
instance-scoped variables were being copied into variables local to
the operator() functions and passed to the worker functions as
individual arguments.

This commit moves all of those copies into structs (well, classes with
all public elements) and then just passes the two structs instead of
the godzillion arguments.

This alone dropped the time to compile BSIM6 from something like an
hour (on my machine, with shared library mode on, which makes
everything compile twice) down to a few minutes.  It dropped the
memory requirement, too.

The BSIM6 still has a bug in pmos sensitivites, so I still have a bug
in the derivative code on this branch.  But now I can debug it with
less time wasted in compilations.

Committing the updated VBIC while I'm at it.  This is still working.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1284